+++ Core/Editor/Inspector.cs
@@ -1,0 +1,11 @@
+using UnityEditor.UIElements;
+using UnityEngine.UIElements;
+
+namespace Core.Editor
+{
+    [UxmlElement]
+    public partial class Inspector : InspectorElement
+    {
+
+    }
+}