+++ Core/Editor/SplitView.cs
@@ -1,0 +1,11 @@
+using UnityEngine.UIElements;
+
+
+namespace Core.Editor
+{
+    [UxmlElement]
+    public partial class SplitView : TwoPaneSplitView
+    {
+        
+    }
+}