+++ StatSystem/Editor/StatCollectionEditor.cs
@@ -1,0 +1,10 @@
+using Core.Editor;
+using UnityEngine.UIElements;
+
+namespace StatSystem.Editor
+{
+    [UxmlElement]
+    public partial class StatCollectionEditor : ScriptableObjectCollectionEditor<StatDefinition>
+    {
+    }
+}