+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 ## [0.0.1] - 2025.11.30
-### Changelog
-- basic folder structure added
 
+### Added
+- Basic stat system script:
+  - Stat, PrimaryStat, Attribute
+  - StatDatabase, StatDefinition, StatModifier
+  - StatController
+- StatDatabase Editor Window
+  - Create new stats
+  - Edit stat names and values
+  - Duplicate existing stats
+  - Delete stats